Paper 6: fix ART-SFT OOM, adapter download, and λ blow-up
First full RunPod ART run surfaced three bugs; all fixed:

- OOM: ART's paired forward held two full 7B autograd graphs live at once
  (normal + ablated), ~2x memory -> 80 GB overflow. ChatLoRATrainer now does
  split backward: backward the SFT(+aux) loss first to free its graph BEFORE
  the ablated forward, then art_loss.backward(); grads accumulate, peak ~one
  graph (control-level). Plus PYTORCH_CUDA_ALLOC_CONF=expandable_segments:True.

- Adapter download: the launcher's `--exclude '*.safetensors'` also dropped the
  LoRA adapter_model.safetensors (the durable artifact). Now excludes the model
  SHARDS by name (model-*.safetensors, model.safetensors) + big-model dirs, so
  adapter weights download.

- λ calibration blow-up: targeting |ART| ~ 10% of L_sft explodes when the gap
  is small (gap 0.023 -> λ 6.29, ART grad ~6x SFT). Capped via max_coefficient
  (default 1.0; λ≈1 balances ART and SFT gradients). Calibration uses a no-grad
  measure_gap().

Also: eval merges base in fp16 (was fp32); remote skips a condition whose
adapter already exists (idempotent re-runs).

Re-run: RUN_DEPENDENCY=0 RUN_ART_SFT=1 RUN_EVAL=1 ./run_session.sh

+++ papers/6_ablation_resistance/scripts/eval_pipeline.py
@@ -85,10 +85,11 @@ def materialize_merged(adapter: str, base_model: str, dest: Path, *, dry_run: bo
     if dry_run:
         logger.info("  [dry-run] would merge %s + adapter %s -> %s", base_model, adapter, dest)
         return str(dest)
+    import torch
     from peft import PeftModel
     from transformers import AutoModelForCausalLM, AutoTokenizer
     logger.info("Merging base %s + adapter %s -> %s", base_model, adapter, dest)
-    base = AutoModelForCausalLM.from_pretrained(base_model)
+    base = AutoModelForCausalLM.from_pretrained(base_model, torch_dtype=torch.float16)
     merged = PeftModel.from_pretrained(base, adapter).merge_and_unload()
     dest.mkdir(parents=True, exist_ok=True)
     merged.save_pretrained(dest)
